feat(ui): extract shared UI framework from claude-history-dashboard
