+++ src/claude-history-dashboard/src/components/ui/badge.tsx
@@ -1,40 +1,1 @@
-import { Slot } from "@radix-ui/react-slot";
-import { cva, type VariantProps } from "class-variance-authority";
-import type * as React from "react";
-
-import { cn } from "@/lib/utils";
-
-const badgeVariants = cva(
-	"inline-flex items-center justify-center rounded-full border px-2 py-0.5 text-xs font-medium w-fit whitespace-nowrap shrink-0 [&>svg]:size-3 gap-1 [&>svg]:pointer-events-none focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px] aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive transition-[color,box-shadow] overflow-hidden",
-	{
-		variants: {
-			variant: {
-				default: "border-transparent bg-primary text-primary-foreground [a&]:hover:bg-primary/90",
-				secondary: "border-transparent bg-secondary text-secondary-foreground [a&]:hover:bg-secondary/90",
-				destructive:
-					"border-transparent bg-destructive text-white [a&]:hover:bg-destructive/90 focus-visible:ring-destructive/20 dark:focus-visible:ring-destructive/40 dark:bg-destructive/60",
-				outline: "text-foreground [a&]:hover:bg-accent [a&]:hover:text-accent-foreground",
-				cyber:
-					"glass-card border-primary/30 text-primary bg-transparent [a&]:hover:border-primary/50 [a&]:hover:bg-primary/10",
-				"cyber-secondary":
-					"glass-card border-secondary/30 text-secondary bg-transparent [a&]:hover:border-secondary/50 [a&]:hover:bg-secondary/10",
-			},
-		},
-		defaultVariants: {
-			variant: "cyber",
-		},
-	}
-);
-
-function Badge({
-	className,
-	variant,
-	asChild = false,
-	...props
-}: React.ComponentProps<"span"> & VariantProps<typeof badgeVariants> & { asChild?: boolean }) {
-	const Comp = asChild ? Slot : "span";
-
-	return <Comp data-slot="badge" className={cn(badgeVariants({ variant }), className)} {...props} />;
-}
-
-export { Badge, badgeVariants };
+export { Badge, badgeVariants } from "../../../../../utils/ui/components/badge";
