+++ src/utils/ui/theme/cyberpunk.css
@@ -1,0 +1,322 @@
+/* Cyberpunk theme styles for authentication pages */
+
+/* Glass card effect */
+.glass-card {
+	background: rgba(10, 10, 20, 0.9);
+	backdrop-filter: blur(20px);
+	-webkit-backdrop-filter: blur(20px);
+	border: 1px solid rgba(255, 149, 0, 0.1);
+}
+
+/* Neon glow effects */
+.neon-glow {
+	box-shadow:
+		0 0 10px rgba(255, 149, 0, 0.3),
+		0 0 20px rgba(255, 149, 0, 0.2),
+		0 0 40px rgba(255, 149, 0, 0.1);
+}
+
+.neon-amber {
+	text-shadow:
+		0 0 10px rgba(255, 149, 0, 0.5),
+		0 0 20px rgba(255, 149, 0, 0.3),
+		0 0 40px rgba(255, 149, 0, 0.2);
+}
+
+.neon-cyan {
+	text-shadow:
+		0 0 10px rgba(0, 240, 255, 0.5),
+		0 0 20px rgba(0, 240, 255, 0.3),
+		0 0 40px rgba(0, 240, 255, 0.2);
+}
+
+/* Neon border */
+.neon-border {
+	border: 1px solid rgba(255, 149, 0, 0.3);
+	box-shadow:
+		inset 0 0 20px rgba(255, 149, 0, 0.05),
+		0 0 20px rgba(255, 149, 0, 0.1);
+}
+
+/* Tech corner decoration */
+.tech-corner::before,
+.tech-corner::after {
+	content: "";
+	position: absolute;
+	width: 20px;
+	height: 20px;
+	border: 2px solid rgba(255, 149, 0, 0.4);
+}
+
+.tech-corner::before {
+	top: -1px;
+	left: -1px;
+	border-right: none;
+	border-bottom: none;
+	border-radius: 10px 0 0 0;
+}
+
+.tech-corner::after {
+	bottom: -1px;
+	right: -1px;
+	border-left: none;
+	border-top: none;
+	border-radius: 0 0 10px 0;
+}
+
+/* Cyberpunk grid background */
+.cyber-grid {
+	background-image:
+		linear-gradient(rgba(255, 149, 0, 0.15) 1px, transparent 1px),
+		linear-gradient(90deg, rgba(255, 149, 0, 0.15) 1px, transparent 1px);
+	background-size: 50px 50px;
+}
+
+/* Scan lines effect */
+.scan-lines {
+	background: repeating-linear-gradient(
+		0deg,
+		rgba(0, 0, 0, 0.4) 0px,
+		rgba(0, 0, 0, 0.4) 1px,
+		transparent 1px,
+		transparent 3px
+	);
+	pointer-events: none;
+}
+
+/* Gradient text */
+.gradient-text {
+	background: linear-gradient(135deg, #ff9500, #ffd700, #00f0ff);
+	-webkit-background-clip: text;
+	-webkit-text-fill-color: transparent;
+	background-clip: text;
+}
+
+/* Animation: Fade in up */
+.animate-fade-in-up {
+	animation: fadeInUp 0.5s ease-out forwards;
+}
+
+@keyframes fadeInUp {
+	from {
+		opacity: 0;
+		transform: translateY(20px);
+	}
+	to {
+		opacity: 1;
+		transform: translateY(0);
+	}
+}
+
+/* Animation delays */
+.delay-100 {
+	animation-delay: 100ms;
+}
+
+.delay-200 {
+	animation-delay: 200ms;
+}
+
+/* Animation: Ripple */
+@keyframes ripple {
+	0% {
+		transform: scale(0.8);
+		opacity: 0.5;
+	}
+	100% {
+		transform: scale(1.2);
+		opacity: 0;
+	}
+}
+
+.animate-ripple {
+	animation: ripple 4s ease-out infinite;
+}
+
+.animate-ripple-delayed {
+	animation: ripple 4s ease-out infinite;
+	animation-delay: 1.3s;
+}
+
+.animate-ripple-delayed-2 {
+	animation: ripple 4s ease-out infinite;
+	animation-delay: 2.6s;
+}
+
+/* Animation: Pulse glow (for running timers) */
+@keyframes pulseGlow {
+	0%,
+	100% {
+		filter: brightness(1);
+	}
+	50% {
+		filter: brightness(1.1);
+	}
+}
+
+.animate-pulse-glow {
+	animation: pulseGlow 2s ease-in-out infinite;
+}
+
+/* Animation: Flash (for countdown completion) */
+@keyframes flash {
+	0%,
+	100% {
+		box-shadow: 0 0 20px rgba(255, 149, 0, 0.3);
+	}
+	50% {
+		box-shadow: 0 0 40px rgba(255, 149, 0, 0.8);
+	}
+}
+
+.animate-flash {
+	animation: flash 0.5s ease-in-out 3;
+}
+
+/* Animation: Flicker */
+@keyframes flicker {
+	0%,
+	100% {
+		opacity: 1;
+	}
+	50% {
+		opacity: 0.8;
+	}
+	75% {
+		opacity: 0.9;
+	}
+}
+
+.animate-flicker {
+	animation: flicker 4s ease-in-out infinite;
+}
+
+/* Button glow effect */
+.btn-glow {
+	position: relative;
+	overflow: hidden;
+}
+
+.btn-glow::after {
+	content: "";
+	position: absolute;
+	top: 0;
+	left: -100%;
+	width: 100%;
+	height: 100%;
+	background: linear-gradient(90deg, transparent, rgba(255, 255, 255, 0.2), transparent);
+	transition: left 0.5s ease;
+}
+
+.btn-glow:hover::after {
+	left: 100%;
+}
+
+/* Custom scrollbar */
+.cyberpunk-scrollbar::-webkit-scrollbar {
+	width: 6px;
+}
+
+.cyberpunk-scrollbar::-webkit-scrollbar-track {
+	background: rgba(0, 0, 0, 0.2);
+	border-radius: 3px;
+}
+
+.cyberpunk-scrollbar::-webkit-scrollbar-thumb {
+	background: rgba(255, 149, 0, 0.5);
+	border-radius: 3px;
+}
+
+.cyberpunk-scrollbar::-webkit-scrollbar-thumb:hover {
+	background: rgba(255, 149, 0, 0.7);
+}
+
+/* Glitch effect for error states */
+@keyframes glitch {
+	0%,
+	100% {
+		transform: translateX(0);
+	}
+	20% {
+		transform: translateX(-2px);
+	}
+	40% {
+		transform: translateX(2px);
+	}
+	60% {
+		transform: translateX(-1px);
+	}
+	80% {
+		transform: translateX(1px);
+	}
+}
+
+.glitch-effect {
+	animation: glitch 0.3s ease-in-out infinite;
+}
+
+/* Slide in from right animation for sidebar */
+@keyframes slide-in-right {
+	from {
+		transform: translateX(100%);
+		opacity: 0;
+	}
+	to {
+		transform: translateX(0);
+		opacity: 1;
+	}
+}
+
+.animate-slide-in-right {
+	animation: slide-in-right 0.3s ease-out forwards;
+}
+
+/* Skeleton shimmer animation - neon amber sweep */
+@keyframes skeleton-shimmer {
+	0% {
+		background-position: -200% 0;
+	}
+	100% {
+		background-position: 200% 0;
+	}
+}
+
+.animate-skeleton-shimmer {
+	animation: skeleton-shimmer 2s ease-in-out infinite;
+}
+
+/* Data stream effect for chart bars */
+@keyframes data-stream {
+	0% {
+		opacity: 0.3;
+		transform: scaleY(0.3);
+	}
+	50% {
+		opacity: 0.6;
+	}
+	100% {
+		opacity: 0.3;
+		transform: scaleY(1);
+	}
+}
+
+.animate-data-stream {
+	animation: data-stream 1.5s ease-in-out infinite;
+}
+
+/* Neon pulse for loading indicators */
+@keyframes neon-pulse {
+	0%,
+	100% {
+		box-shadow: 0 0 5px rgba(0, 240, 255, 0.3);
+	}
+	50% {
+		box-shadow:
+			0 0 15px rgba(0, 240, 255, 0.6),
+			0 0 30px rgba(0, 240, 255, 0.3);
+	}
+}
+
+.animate-neon-pulse {
+	animation: neon-pulse 1.5s ease-in-out infinite;
+}
